feat - criado crud de usuario com testes unitarios

+++ src/main/java/br/com/LeiloaBoardgames/service/UsuarioService.java
@@ -3,9 +3,11 @@ package br.com.LeiloaBoardgames.service;
 import br.com.LeiloaBoardgames.domain.Usuario;
 import br.com.LeiloaBoardgames.exceptions.BusinessException;
 import br.com.LeiloaBoardgames.repository.UsuarioRepository;
-import javassist.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.NoSuchElementException;
 
 @RequiredArgsConstructor
 @Service
@@ -19,7 +21,7 @@ public class UsuarioService {
     }
 
 
-    public Usuario atualizar(Integer id, Usuario usuario) throws NotFoundException {
+    public Usuario atualizar(Integer id, Usuario usuario) throws NoSuchElementException {
         Usuario usuarioAtual = buscarPorId(id);
 
         validarUsuarioExistente(usuario);
@@ -31,25 +33,33 @@ public class UsuarioService {
     }
 
 
-
-    public void deletar(Integer id) throws NotFoundException {
+    public Integer deletar(Integer id) throws NoSuchElementException {
         Usuario usuarioExistente = buscarPorId(id);
+        if (!usuarioExistente.getAtivo()) {
+            throw new BusinessException("Usuário já deletado");
+        }
         repository.delete(usuarioExistente);
+        return usuarioExistente.getIdUsuario();
     }
 
-    public Usuario buscarPorId(Integer id) throws NotFoundException {
-        return repository.findById(id).orElseThrow(() -> new NotFoundException("Usuário não encontrado"));
+
+    public Usuario buscarPorId(Integer id) throws NoSuchElementException {
+        return repository.findById(id).orElseThrow(() -> new NoSuchElementException("Usuário não encontrado"));
     }
 
     private void validarUsuarioExistente(Usuario usuario) {
-        if (repository.existsByEmail(usuario.getEmail())) {
+        if (repository.existsByEmail(usuario.getEmail()) && usuario.getAtivo()) {
             throw new BusinessException("Email já cadastrado");
         }
-        if (repository.existsByCpf(usuario.getCpf())) {
+        if (repository.existsByCpf(usuario.getCpf()) && usuario.getAtivo()) {
             throw new BusinessException("CPF já cadastrado");
         }
-        if (repository.existsByUsuario(usuario.getUsuario())) {
+        if (repository.existsByUsuario(usuario.getUsuario()) && usuario.getAtivo()) {
             throw new BusinessException("Usuário já cadastrado");
         }
+    }
+
+    public List<Usuario> buscarTodos() {
+        return repository.findAll();
     }
 }
